ex02 finished

+++ cpp-module-02/ex01/main.cpp
@@ -6,22 +6,22 @@
 /*   By: roguigna <roguigna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/24 14:37:16 by roguigna          #+#    #+#             */
-/*   Updated: 2024/07/27 16:49:54 by roguigna         ###   ########.fr       */
+/*   Updated: 2024/07/29 12:00:14 by roguigna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Fixed.hpp"
 
-	int main( void ) {
+int main( void ) {
 	Fixed a;
 	Fixed const b( 10 );
 	Fixed const c( 42.42f );
 	Fixed const d( b );
 	a = Fixed( 1234.4321f );
-	std::cout << "a is " << a << std::endl;
-	std::cout << "b is " << b << std::endl;
-	std::cout << "c is " << c << std::endl;
-	std::cout << "d is " << d << std::endl;
+	std::cout << "a is " << a.toFloat() << std::endl;
+	std::cout << "b is " << b.toFloat() << std::endl;
+	std::cout << "c is " << c.toFloat() << std::endl;
+	std::cout << "d is " << d.toFloat() << std::endl;
 	std::cout << "a is " << a.toInt() << " as integer" << std::endl;
 	std::cout << "b is " << b.toInt() << " as integer" << std::endl;
 	std::cout << "c is " << c.toInt() << " as integer" << std::endl;
